Prevent attempts to modify tables defined as queries
Previously this may have worked if the foreign table name had the same
name as an existing table on the Firebird server, but that would have
been at the "happy coincidence" level of database operations.

Should, for whatever reason, a table defined as a query need to
have DML operations performed on it, these can be done via triggers
which operate on a correctly defined foreign table referencing the
actual table on the Firebird server.