IMPORT SCHEMA: add "quote_identifier" option where appropriate
If importing a Firebird table where the table name or column name is
all-lower case, PostgreSQL won't know it needs quoting when being passed
to Firebird, so we'll need to explicitly add the "quote_identifier".

Note this currently only checks if the first character of the name
is lower case.