Push down boolean tests to Firebird
Where possible, push down boolean tests (IS TRUE, IS NOT FALSE etc.) to
Firebird.

Note that Firebird's interpretation of "IS NOT TRUE" and "IS NOT FALSE"
is different to that of PostgreSQL (it ignores NULL values) so the
pushdown query needs to be constructed so it returns the same results
as a native PostgreSQL query would.

Firebird 3.0 and later.