Add experimental support for implicit boolean columns
This experimental patch makes it possible to map a PostgreSQL boolean
column to a Firebird integer or numeric column of any type representing an
implicit boolean value. This makes it possible to query the column as if
it were a true boolean.

Addresses GitHub #9.

Configuration
-------------

This functionality is disabled by default, and must be enabled by setting
the server-level option "implicit_bool_type" to "true".

For Firebird 2.5.x no further configuration is needed, and firebird_fdw
will automatically assume that a boolean column in a foreign table definition
maps to an implicit boolean column in the Firebird table.

With Firebird 3 and later, by default firebird_fdw assumes a boolean
column in a foreign table definition maps to a boolean column in the Firebird
table. To indicate that the Firebird table column is an implicit boolean, set
the column-level option "implicit_bool_type" to "true".

Usage
-----

firebird_fdw performs following mapping of implicit boolean values:

 - 0 is assumed to represent "false"
 - any other value is assumed to represent "true"
 - NULL remains NULL

When inserting data into an implicit boolean column firebird_fdw
performs following mapping:

 - "false" is translated to 0
 - "true" is translated to 1
 - NULL remains NULL

Queries with implicit boolean tests will be pushed down to Firebird in
exactly the same way as queries with true boolean tests.

Example
-------

Firebird:

    CREATE TABLE booltest (
      id INT NOT NULL,
      boolval SMALLINT
    );

    INSERT INTO booltest VALUES(1, 0);
    INSERT INTO booltest VALUES(2, 1);
    INSERT INTO booltest VALUES(3, 2);
    INSERT INTO booltest VALUES(4, NULL);

PostgreSQL:

    ALTER SERVER fb_test OPTIONS (implicit_bool_type 'true');

    CREATE FOREIGN TABLE fb_booltest (
      id INT NOT NULL,
      boolval BOOL OPTIONS (implicit_bool_type 'true')
    )
    SERVER fb_test
    OPTIONS (table_name 'booltest');

    SELECT * FROM fb_booltest;
     id | boolval
    ----+---------
      1 | f
      2 | t
      3 | t
      4 |
    (4 rows)

     SELECT * FROM fb_booltest WHERE boolval;
     id | boolval
    ----+---------
      2 | t
      3 | t

    EXPLAIN SELECT * FROM fb_booltest WHERE boolval;
                                      QUERY PLAN
    ------------------------------------------------------------------------------
     Foreign Scan on fb_booltest  (cost=10.00..14.00 rows=4 width=5)
       Firebird query: SELECT id, boolval <> 0 FROM booltest WHERE (boolval <> 0)
    (2 rows)

    SELECT * FROM fb_booltest WHERE boolval IS NOT TRUE;
     id | boolval
    ----+---------
      1 | f
      4 |
    (2 rows)

    EXPLAIN SELECT * FROM fb_booltest WHERE boolval IS NOT TRUE;
                                                 QUERY PLAN
    ----------------------------------------------------------------------------------------------------
     Foreign Scan on fb_booltest  (cost=10.00..14.00 rows=4 width=5)
       Firebird query: SELECT id, boolval <> 0 FROM booltest WHERE ((boolval = 0) OR (boolval IS NULL))
    (2 rows)

Limitations
-----------

This patch is experimental and it's possible there are circumstances
when it may generate invalid Firebird SQL due to unanticipated query
tree variants.