IMPORT SCHEMA: further fix handling of identifiers in filter clauses
It turns out that the PostgreSQL FDW API silently rejects table
definitions which do not exactly match the table names specified in the
"LIMIT TO" clause.

E.g. if "LIMIT TO(foo)" is specified, but firebird_fdw returns
"CREATE TABLE public.FOO", the table defintion will be silently
dropped. See "IsImportableForeignTable()" in
src/backend/foreign/foreign.c.

To work around this, we create a rather inelegant CTE in the Firebird
metadata query to pass through the table names as provided in
"LIMIT TO" and as generated for the Firebird metadata query. That
makes it possible to select the "correct" rendering of the table name.