Makefile: further cleanup
The PGXS TAP_TESTS option was added in PostgreSQL 12, but this extension
currently supports them from 9.5, so we can't really make use of that
yet.

However, in the interests of future compatibility, rename the exported
VERSION_NUM environment variable to PG_VERSION_NUM. This will make it
possible to export it directly from the Makefile should we ever reach
a point so far in the future we can enable the TAP_TESTS option and
remove the custom installcheck targets.