Initial support for batch inserts
API support for this functionality was added in PostgreSQL 14 (core
commit b663a41363).

This commit adds the "batch_size" server and table option. If set to
a value greater than 1, firebird_fdw will use a "prepared" statement
when inserting up to the specified number of rows. This means the
statement only needs to be prepared once for each set of inserts,
reducing overhead.

This is a provisional implementation and may be modified in the future.

Requires:
 - libfq 0.5.0 and later
 - PostgreSQL 14 and later.