Initial support for TRUNCATE
Although Firebird does not currently provide a TRUNCATE command, we
can simulate it with a simple unqualified DELETE operation.

Given this limitation, the main advantage in using TRUNCATE is that
multiple tables can be specified in a single command.

Note however that with thit initial implementation there is no support
for TRUNCATE ... CASCADE or for truncating tables with foreign key
references from other Firebird tables (even if all dependent tables
are specified in the TRUNCATE command).