Addi basic support for UUID datatype
Firebird does not have an explicit UUID type, but provides a function
"gen_uuid()" which generates a UUID as a series of raw byte values.
These can either be stored in a column defined with "CHARACTER SET
OCTETS", or converted to a string with "uuid_to_char()" for storage
in a normal text column.

However, if a PostgreSQL UUID value is pushed down as a string
literal, any UUID comparisons will fail unless the corresponding
value in Firebird is normalised in the same way as PostgreSQL
(which is likely not the case, as "uuid_to_char()" produces
UUIDs formatted with upper-case characters, whereas PostgreSQL
normalizes to lower-case values, per RFC4122 [1]).

To avoid this, we can simply not push down UUID values, so
comparisons can be performed on the PostgreSQL side, which
accepts a range of formats as valid UUID input values [2].

This is less efficient than performing comparisons on the Firebird
side; this could be implemented later, but would require knowledge of
how the UUID is stored in Firebird (expressed as a column-level
option or similar).

Note that currently no provision exists for inserting UUID
values into Firebird "CHARACTER SET OCTETS" columns.

[1] https://datatracker.ietf.org/doc/html/rfc4122
[2] https://www.postgresql.org/docs/current/datatype-uuid.html

This patch assumes a libfq version with support for
"CHARACTER SET OCTETS".

Addresses issues in GitHub #42.