Fix generated columns support
Support for generated columns (PostgreSQL 12 and later) was added in
the mistaken assumption that the generated column value is generated
in PostgreSQL and passed to Firebird. However, the foreign table
definition should actually be a representation of the remote table
(as is the case e.g. with NOT NULL), and the value should be generated
by Firebird.

Happily Firebird has supported generated columns since at least 2.5,
so this isn't an issue. Note that Firebird uses the syntax
"COMPUTED BY" by default, but also accepts "GENERATED ALWAYS AS"

This does mean that any users of generated columns should check their
setup before upgrading to firebird_fdw 1.4.0.